pyproject.toml
[project]
name = "aidevs3"
version = "0.1.0"
description = "AI_devs3 course"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Piotr Święcik", email = "piotr.swiecik@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Repository = "https://github.com/piotrswiecik/aidevs3"

[tool.poetry]
name = "aidevs3"
version = "0.1.0"
description = "AI_devs3 course"
authors = ["Piotr Święcik <piotr.swiecik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
openai = "^1.54.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.10.10"
langfuse = "^2.53.9"
tiktoken = "^0.8.0"
fastapi = {extras = ["standard"], version = "^0.115.4"}
faiss-cpu = "^1.9.0"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
    "integration",
]
log_cli = "True"
log_cli_level = "INFO"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
